Fix for broken coop test
For some reason, the engineer that is tracked by the camera isn't being
created. It works on a local version of coop but not through FAF. I have
no idea why as the script is fine. This should fix the issue until we
figure out what's causing the error. I have no idea why this is a
problem.